Added support for using the LKH as a local search (requires external executable, does not yet test for it's presence)
Former-commit-id: aff934e1739aee511c99391719202be691c65d86

+++ memetico/src/java/memetico/logging/IPCLogger.java
@@ -5,7 +5,7 @@ import org.marcos.uon.tspaidemo.util.log.ILogger;
 
 public interface IPCLogger extends ILogger<MemeticoSnapshot> {
     interface View extends ILogger.View<MemeticoSnapshot> {
-        double getStartTime();
+        long getStartTime();
     }
 
     /**
